+++ deltafi-common/src/main/java/org/deltafi/common/types/ActionEventType.java
@@ -20,6 +20,7 @@ package org.deltafi.common.types;
 public enum ActionEventType {
   TRANSFORM,
   LOAD,
+  DOMAIN,
   ENRICH,
   FORMAT,
   FORMAT_MANY,