+++ deltafi-common/src/main/java/org/deltafi/common/types/DomainResultInput.java
@@ -22,18 +22,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
+import java.util.Map;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(builderMethodName = "newBuilder")
-public class ActionRegistrationInput {
-  private List<TransformActionSchemaInput> transformActions;
-  private List<LoadActionSchemaInput> loadActions;
-  private List<DomainActionSchemaInput> domainActions;
-  private List<EnrichActionSchemaInput> enrichActions;
-  private List<FormatActionSchemaInput> formatActions;
-  private List<ValidateActionSchemaInput> validateActions;
-  private List<EgressActionSchemaInput> egressActions;
+public class DomainResultInput {
+  private Map<String, String> indexedMetadata;
 }